+++ default_back.json
@@ -1,0 +1,17 @@
+{
+  "pixelsPerInch": 300,
+  "paperSize": [8.5, 11],
+  "rows": 5,
+  "columns": 2,
+  "borderWidth": 40,
+  "logoMargin": 40,
+  "topMargin": 0.5,
+  "sideMargin": 0.75,
+  "textMargin": 30,
+  "backgroundColor": "white",
+  "logoPath": "fair_logo.jpg",
+  "logoSize": [369, 273],
+  "resourcePath": "..//resource//",
+  "font": "times.ttf",
+  "defaultData": "fair-data.csv"
+}
