--- saveload.py
@@ -1,6 +1,0 @@
-import json
-
-def load_config(filePath):
-    file = open(filePath, 'r')
-    file = json.load(file)
-    return file
